+++ projects/plugins/boost/app/modules/optimizations/lcp/class-lcp.php
@@ -23,30 +23,28 @@ class Lcp implements Feature, Changes_Output_After_Activation, Optimization, Has
 	const TYPE_IMAGE = 'img';
 
 	/**
-	 * LCP storage class instance.
+	 * The LCP data of the current request.
 	 *
-	 * @var LCP_Storage
+	 * @var array|false
 	 */
-	private $storage;
-
-	/**
-	 * Output filter instance.
-	 *
-	 * @var Output_Filter
-	 */
-	private $output_filter;
+	private $lcp_data;
 
 	public function setup() {
-		$this->output_filter = new Output_Filter();
-		$this->storage       = new LCP_Storage();
-
+		add_action( 'wp', array( $this, 'on_wp_load' ), 1 );
 		add_action( 'template_redirect', array( $this, 'add_output_filter' ), -999999 );
+
 		add_action( 'jetpack_boost_lcp_invalidated', array( $this, 'handle_lcp_invalidated' ) );
 
-		// Initialize the optimizer for background images. Doing it late enough so wp can load, but before any output is sent.
-		add_action( 'wp', array( LCP_Optimize_Bg_Image::class, 'init' ) );
+		// Skip images optimized LCP images from being processed by image-cdn.
+		add_filter( 'jetpack_photon_skip_image', array( $this, 'skip_cdn_image' ), 10, 2 );
 
 		LCP_Invalidator::init();
+	}
+
+	public function on_wp_load() {
+		$this->lcp_data = ( new LCP_Storage() )->get_current_request_lcp();
+
+		LCP_Optimize_Bg_Image::init( $this->lcp_data );
 	}
 
 	public function add_output_filter() {
@@ -54,7 +52,32 @@ class Lcp implements Feature, Changes_Output_After_Activation, Optimization, Has
 			return;
 		}
 
-		$this->output_filter->add_callback( array( $this, 'optimize' ) );
+		$output_filter = new Output_Filter();
+		$output_filter->add_callback( array( $this, 'optimize_lcp_img_tag' ) );
+	}
+
+	/**
+	 * Filter to skip images optimized LCP images from being processed by image-cdn.
+	 *
+	 * If image-cdn is processing the image, it will change the markup of the tag and we will not be able to find the tag while trying to apply LCP optimization.
+	 *
+	 * @param bool   $skip Whether to skip the image.
+	 * @param string $image_url The image URL.
+	 *
+	 * @return bool Whether to skip the image.
+	 */
+	public function skip_cdn_image( $skip, $image_url ) {
+		if ( empty( $this->lcp_data ) ) {
+			return $skip;
+		}
+
+		foreach ( $this->lcp_data as $lcp_element ) {
+			if ( $lcp_element['type'] === self::TYPE_IMAGE && $lcp_element['url'] === $image_url ) {
+				return true;
+			}
+		}
+
+		return $skip;
 	}
 
 	/**
@@ -67,18 +90,16 @@ class Lcp implements Feature, Changes_Output_After_Activation, Optimization, Has
 	 *
 	 * @since 3.13.1
 	 */
-	public function optimize( $buffer_start, $buffer_end ) {
-		$lcp_storage = $this->storage->get_current_request_lcp();
-
-		if ( empty( $lcp_storage ) ) {
+	public function optimize_lcp_img_tag( $buffer_start, $buffer_end ) {
+		if ( empty( $this->lcp_data ) ) {
 			return array( $buffer_start, $buffer_end );
 		}
 
 		// Combine the buffers for processing
 		$combined_buffer = $buffer_start . $buffer_end;
 
-		foreach ( $lcp_storage as $lcp_data ) {
-			$optimizer = new LCP_Optimize_Img_Tag( $lcp_data );
+		foreach ( $this->lcp_data as $lcp_element ) {
+			$optimizer = new LCP_Optimize_Img_Tag( $lcp_element );
 
 			$combined_buffer = $optimizer->optimize_buffer( $combined_buffer );
 		}